+++ pages/changelog.md
@@ -7,7 +7,7 @@ sidebar: auto
 ## v2.0.0a16
 
 - 新增 `MessageFormatter` 可用于 `Message` 的模板生成
-- 新增 `matcher.got` 支持 `MessageFormatter`
+- 新增 `matcher.got` `matcher.send` `matcher.pause` `matcher.reject` `matcher.finish` 支持 `MessageFormatter`
 - 移除 `matcher.got` 原本的 `state format` 支持，由 `MessageFormatter` template 替代
 - `adapter` 基类拆分为单独文件
 - 修复 `fastapi` Driver Websocket 未能正确提供请求头部