+++ pages/changelog.md
@@ -11,6 +11,7 @@ sidebar: auto
 - 移除 `matcher.got` 原本的 `state format` 支持，由 `MessageFormatter` template 替代
 - `adapter` 基类拆分为单独文件
 - 修复 `fastapi` Driver Websocket 未能正确提供请求头部
+- 新增 `fastapi` Driver 更多的 uvicorn 相关配置项
 
 ## v2.0.0a15